Bootstrap: Don't download twice

+++ bootstrap/src/main/java/net/cubition/bootstrap/Resource.java
@@ -7,6 +7,8 @@ import org.apache.commons.io.IOUtils;
 import java.io.*;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collections;
 
@@ -213,6 +215,7 @@ public class Resource implements Serializable {
 
         // Open a basic stream, if possible
         URL jsonURL = new URL(remotePath + ".json");
+        System.out.println("Grabbing " + jsonURL.toString());
         try (InputStream jsonIn = new BufferedInputStream(jsonURL.openStream());
              OutputStream jsonOut = new FileOutputStream(this.localPath + ".json")) {
             IOUtils.copy(jsonIn, jsonOut);
@@ -227,12 +230,15 @@ public class Resource implements Serializable {
         JsonObject description = new Gson().fromJson(contents, JsonObject.class);
 
         // Grab the extension from this description
-        String extension = description.has("type") ? description.get("type").getAsString() : "jar";
+        String extension = description.has("type") ? "." + description.get("type").getAsString() : "jar";
 
         // Download the actual Resource now we have the info we need
         this.localPath += extension;
         remotePath += extension;
 
+        if (Files.exists(Paths.get(localPath))) return true;
+
+        System.out.println("Grabbing " + remotePath);
         try (FileOutputStream localOut = new FileOutputStream(this.localPath);
              InputStream content = new BufferedInputStream(new URL(remotePath).openStream())) {
             IOUtils.copy(content, localOut);
